backend/app/ai_tutor_core.py
import torch
import torch.nn as nn
from transformers import AutoModelForCausalLM, AutoTokenizer
from sklearn.manifold import TSNE
from sklearn.preprocessing import StandardScaler
import numpy as np
from typing import Dict, List, Optional, Tuple, Union
from dataclasses import dataclass
from enum import Enum
import json
import os
from dotenv import load_dotenv
import asyncio

class LearningStyle(Enum):
    VISUAL = "visual"
    AUDITORY = "auditory"
    KINESTHETIC = "kinesthetic"
    READING_WRITING = "reading_writing"

@dataclass
class StudentProfile:
    learning_style: LearningStyle
    comprehension_level: float
    attention_span: float
    preferred_difficulty: str
    recent_mistakes: List[str]
    mastered_concepts: List[str]
    cognitive_load: float
    emotional_state: str
    engagement_score: float

class AdvancedMathTutorAI:
    def __init__(self, use_gpu: bool = True):
        load_dotenv()
        self.device = torch.device("cuda" if torch.cuda.is_available() and use_gpu else "cpu")
        self._initialize_models()
        self.teaching_strategies = self._load_teaching_strategies()
        self.learning_adaptations = self._initialize_learning_adaptations()
        self.performance_metrics = self._initialize_performance_metrics()
        self.knowledge_graph = self._initialize_knowledge_graph()
        self.visualization_engine = self._initialize_visualization_engine()

    def _initialize_models(self):
        # Main tutoring model
        self.main_model = self._load_model("mistralai/Mistral-7B-Instruct-v0.1")
        
        # Specialized models
        self.models = {
            "explanation": self._load_model(os.getenv("EXPLANATION_MODEL_PATH")),
            "assessment": self._load_model(os.getenv("ASSESSMENT_MODEL_PATH")),
            "hint_generation": self._load_model(os.getenv("HINT_MODEL_PATH")),
            "engagement": self._load_model(os.getenv("ENGAGEMENT_MODEL_PATH"))
        }
        
        # Initialize tokenizers
        self.tokenizer = AutoTokenizer.from_pretrained(
            "mistralai/Mistral-7B-Instruct-v0.1",
            use_fast=True
        )

    def _load_model(self, model_path: str) -> Optional[nn.Module]:
        try:
            model = AutoModelForCausalLM.from_pretrained(model_path).to(self.device)
            
            # Apply optimizations
            if self.device.type == "cuda":
                model = torch.quantization.quantize_dynamic(
                    model, {torch.nn.Linear}, dtype=torch.qint8
                )
            return model
        except Exception as e:
            print(f"Error loading model {model_path}: {e}")
            return None

    async def generate_explanation(
        self,
        concept: str,
        student_profile: StudentProfile,
        previous_explanations: Optional[List[str]] = None
    ) -> Dict[str, any]:
        """Generate personalized concept explanation"""
        # Adapt to learning style
        adapted_content = self._adapt_to_learning_style(
            concept,
            student_profile.learning_style
        )
        
        # Generate explanation using specialized model
        explanation = await self._generate_specialized_content(
            "explanation",
            adapted_content,
            student_profile
        )
        
        # Add visual aids if needed
        visual_aids = self._generate_visual_aids(concept) if student_profile.learning_style == LearningStyle.VISUAL else None
        
        return {
            "explanation": explanation,
            "visual_aids": visual_aids,
            "difficulty_level": self._calculate_explanation_difficulty(explanation),
            "recommended_practice": self._generate_practice_recommendations(concept, student_profile)
        }

    async def assess_solution(
        self,
        problem: str,
        student_solution: str,
        correct_solution: str,
        student_profile: StudentProfile
    ) -> Dict[str, any]:
        """Assess student's solution comprehensively"""
        # Analyze solution using specialized model
        assessment = await self._generate_specialized_content(
            "assessment",
            {
                "problem": problem,
                "student_solution": student_solution,
                "correct_solution": correct_solution
            },
            student_profile
        )
        
        # Calculate metrics
        similarity_score = self._calculate_solution_similarity(
            student_solution,
            correct_solution
        )
        
        # Generate feedback
        feedback = self._generate_personalized_feedback(
            assessment,
            student_profile
        )
        
        return {
            "is_correct": similarity_score > 0.9,
            "similarity_score": similarity_score,
            "detailed_feedback": feedback,
            "misconceptions": self._identify_misconceptions(assessment),
            "recommended_practice": self._generate_practice_recommendations(
                problem,
                student_profile,
                assessment
            )
        }

    async def provide_hint(
        self,
        problem: str,
        student_profile: StudentProfile,
        previous_hints: Optional[List[str]] = None
    ) -> Dict[str, any]:
        """Provide personalized hints"""
        # Generate hint using specialized model
        hint = await self._generate_specialized_content(
            "hint_generation",
            {
                "problem": problem,
                "previous_hints": previous_hints or []
            },
            student_profile
        )
        
        return {
            "hint": hint,
            "hint_level": len(previous_hints) + 1 if previous_hints else 1,
            "visual_aid": self._generate_hint_visualization(hint) if student_profile.learning_style == LearningStyle.VISUAL else None
        }

    async def track_engagement(
        self,
        student_profile: StudentProfile,
        session_data: Dict
    ) -> Dict[str, float]:
        """Track student engagement in real-time"""
        engagement_metrics = await self._analyze_engagement(
            session_data,
            student_profile
        )
        
        return {
            "attention_level": engagement_metrics["attention"],
            "interaction_rate": engagement_metrics["interaction"],
            "emotional_state": engagement_metrics["emotion"],
            "cognitive_load": engagement_metrics["cognitive_load"]
        }

    def optimize_learning_path(
        self,
        student_profile: StudentProfile,
        learning_goals: List[str]
    ) -> Dict[str, any]:
        """Generate optimized learning path"""
        current_knowledge = self._assess_current_knowledge(student_profile)
        path = self._generate_learning_path(current_knowledge, learning_goals)
        
        return {
            "learning_path": path,
            "estimated_completion_time": self._estimate_completion_time(path),
            "prerequisites": self._identify_prerequisites(path),
            "milestones": self._generate_milestones(path)
        }

    def _adapt_to_learning_style(
        self,
        content: str,
        style: LearningStyle
    ) -> str:
        """Adapt content to student's learning style"""
        adaptation = self.learning_adaptations.get(style, {})
        return adaptation.get("adapt_func", lambda x: x)(content)

    async def _generate_specialized_content(
        self,
        model_type: str,
        content: Dict,
        student_profile: StudentProfile
    ) -> str:
        """Generate content using specialized model"""
        if model_type not in self.models:
            return None
            
        model = self.models[model_type]
        adapted_content = self._adapt_to_learning_style(
            str(content),
            student_profile.learning_style
        )
        
        inputs = self.tokenizer(adapted_content, return_tensors="pt")
        outputs = model.generate(**inputs, max_length=512)
        return self.tokenizer.decode(outputs[0], skip_special_tokens=True)

if __name__ == "__main__":
    # Initialize tutor
    tutor = AdvancedMathTutorAI()
    
    # Create sample student profile
    student_profile = StudentProfile(
        learning_style=LearningStyle.VISUAL,
        comprehension_level=0.7,
        attention_span=20.0,
        preferred_difficulty="medium",
        recent_mistakes=[],
        mastered_concepts=["basic_addition"],
        cognitive_load=0.5,
        emotional_state="focused",
        engagement_score=0.8
    )
    
    # Test explanation generation
    async def test_tutor():
        explanation = await tutor.generate_explanation(
            "multiplication",
            student_profile
        )
        print("Generated explanation successfully")
    
    asyncio.run(test_tutor())