backend/requirements.txt
transformers
torch
plotly
scikit-learn
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
opencv-python
numpy
Pillow
matplotlib
tqdm
